Add blog post about DevOps detachment
A blog post describing the recent change to the DevOps repository has
been written. This describes mostly the importance and relevance of
AAROC/DevOps. I go into some depth about how we use it and why we had
to detach it from the original forks.

Some new project releases co-incide with the  blog post.